✨ Implementation of saving post to User posts array, update package.json

+++ graphql/resolvers/Post/index.js
@@ -2,6 +2,8 @@ import User from "../../../server/models/User";
 import Post from "../../../server/models/Post";
 import Comment from "../../../server/models/Comment";
 
+import { transformPost } from "../merge";
+
 export default {
   Query: {
     post: async (parent, { _id }, context, info) => {
@@ -30,12 +32,27 @@ export default {
         published: post.published,
         author: post.author
       });
-
-      return new Promise((resolve, reject) => {
-        newPost.save((err, res) => {
-          err ? reject(err) : resolve(res);
+      let createdPost;
+      try {
+        // const result = await newPost.save();
+        const result = await new Promise((resolve, reject) => {
+         newPost.save((err, res) => {
+            err ? reject(err) : resolve(res);
+          });
         });
-      });
+        createdPost = transformPost(result);
+        const creator = await User.findById(post.author);
+
+        if (!creator) {
+          throw new Error("User not found.");
+        }
+        creator.posts.push(newPost);
+        await creator.save();
+        return createdPost;
+      } catch (error) {
+        console.log(error);
+        throw error;
+      }
     },
     updatePost: async (parent, { _id, post }, context, info) => {
       return new Promise((resolve, reject) => {
@@ -46,12 +63,28 @@ export default {
         );
       });
     },
-    deletePost: (parent, { _id }, context, info) => {
-      return new Promise((resolve, reject) => {
-        Post.findByIdAndDelete(_id).exec((err, res) => {
-          err ? reject(err) : resolve(res);
+    deletePost: async (parent, { _id }, context, info) => {
+      try {
+        // searching for creator of the post and deleting it from the list
+        const post = await Post.findById(_id);
+        const creator = await User.findById(post.author);
+        if (!creator) {
+          throw new Error("user not found.");
+        }
+        const index = creator.posts.indexOf(_id);
+        if (index > -1) {
+          creator.posts.splice(index, 1);
+        }
+        await creator.save();
+        return new Promise((resolve, reject) => {
+          Post.findByIdAndDelete(_id).exec((err, res) => {
+            err ? reject(err) : resolve(res);
+          });
         });
-      });
+      } catch (error) {
+        console.log(error);
+        throw error;
+      }
     }
   },
   Subscription: {
